Organize crypto-related functionalities into `crypto` folder
